Bug 563: Reset backflipping when Tux spawns.
There is one player object per sector. If Tux backflips into a script trigger
and spawns in another sector, the player object in the first sector will still
have "backflipping = true", and Tux will be backflipping when he returns to the
first sector. This looks wrong and can get the user stuck in "Find the Bigger
Fish". The flag needs to be reset when Tux spawns.

Resolves #563. Thanks to Matt McCutchen for this patch.

SVN-Revision: 6300