supertux/{collision.[ch]pp,sector.cpp}: Store velocities in the constraints.
Assume Tux is falling down with speed 2 and onto a downward moving tilemap with
speed 1. With the current constraints, his vertical motion will be set to zero,
possibly resulting in a hopping motion.

With normal, solid tiles this reset cannot be observed due to the way ground
movement is handled. With unisolid tiles, this results in said hopping.

This commit should be a non-functional change, i.e. the implementation differs
but the overall functionality should be unchanged.

SVN-Revision: 6389