Unstable tile: Dissolve, then disappear without falling down.
The new behavior is as follows: If the sprite has a "dissolve" action, it is
triggered on contact. If the "dissolve" animation ends, or if the sprite
doesn't have a "dissolve" action, the "fall-down" action is triggered. If that
animation is done, or if the sprite doesn't have a "fall-down" action, the
object is removed.

Possible future improvements:
- Make the tile non-solid at the end of "dissolve", rather than at the
  beginning of "fall-down". This way falling only sprites would still be solid.
  For this to be of use the collision detection needs to be improved, though.
- Also, remove the object after is has hit a solid object or has fallen out of
  the screen. The "remove at end of animation" behavior is more of a dirty
  hack.

SVN-Revision: 6479