Created a new definition for Particles.
Previously, the parameters used to define particles were not very intuitive. Specifically, the angles and initial velocity were very muddled. It appeared that the angle parameters were being used to randomize both angle and velocity, but in some bizarre manner.  The new definition allows the bounds for both the actual angle of trajectory and magnitude of the initial velocity to be defined, and a random value between those bounds is what is used for each particle.

In a nutshell: the new definition uses trigonometry in a physically correct and intuitive manner.

Since the old definition is called in a manner that makes particles which look good, it has not yet been replaced. However, these calls should probably be reworked under the new definition to produce the same behavior before removing the old definition.