Scripting: Save data through restarts
Currently scripts can only store data by modifying the savegamefile directly.
Now the functions store and load allow scripts to easily preserve data.
Syntax: store(varname,value), where varname is a string, and value is a
serializable object. A squirrel object is serializible iff it is of type
Array, Table, String, Int, Bool, Float. Arrays and Tables are also serializable,
iff they only contain serializible items (eg. Array of Array of Int is serializable).
To load an object the function load(varname) is provided.

Serialization
Is currently done in the form of S-Expressions.  The file is in the same
directory as the savegame file and has the extension .sss

Implementation
When a object is stored it is tranformed into a internal data structure
representing arrays (sqarray in sqarrayser.cpp) , tables (sqdict in sqdictser.cpp)
 and scalar values (stored using boost::any container). The dictionary class
provides a general storage of names and the associated item in the internal
data structure. When a worldmap is loaded a new dictionary is created,
when a worldmap is left a dictionary is destroyed, and the file is created.

The internal data structures for arrays and tables  are implemented as classes,
which all provide the following methods
  + load(ReaderCollection& r)
    Used for loading from the file
  + get(HSQUIRRELVM v)
    Used for pushing onto the stack
  + serialize(Writer* serialize)
    Used for serializing the object
  + loadArray(HSQUIRRELVM v) (or loadDict)
    Used for loading from squirrel

Issues / Further Modifications

-Currently only basic types are storable. Extend it to "special" C++
 Classes which e.g. implement an Interface Storable or use an auto-
 matically generated wrapper for serializing some classes.

-Add option to store from the game and make such stored items "private".

Closes #530