WillOWisp: get name property
This fixes an issue where the WillOWisp objects couldn't be accessed from the
scripting interface anymore. This happened because their name property was not
correctly retrieved from the level. As such the engine would not expose the
objects to the scripting interface. That caused broken scripts in some levels.
The issue was introduced in commit SuperTux/supertux@19b951f.

Fixes #529.