EditorLevelMenu: use -2 as value to check for OK
This is an attempt to fix an issue where certain values could not be changed in
the level menu. It happened because c8c330a93b89f0bf3b67bb3cd9138004aa7a3435
introduced a check for certain values upon clicking "OK". However, it did that
by checking for clicks on menu items with id = -1, which was also used for
certain other menu items. That caused the pop up menu of those menu items to
disappear immediately upon attempting to enter them.

Closes SuperTux/supertux#577 ("Editor: changing a worldmap's tileset does not
work")