Fix Squirrel Level Edit Bug
Using the function Level.edit(false) through a switch when in
edit mode currently causes the game to crash. This is due to
the current Level object being destroyed when the edit mode is
left, but squirrel is still referencing / trying to use it, and
this leads to the seg fault.

The Fix is, that the GameSession object keeps a reference to the
last delteted level, as an extra member variable, thus squirrel
can still access the level object.

Fixes #207