Add ultra-violet lighting
...because penguins are tetrachromatic
It adds an extra light chanel that enhances the lighting
a bit. There are now two lightmaps: the visible lightmap
and a hidden lightmap. The visible lightmap stores the
red, real value of green chanel and blue. And the hidden
lightmap stores the values of real red, ultra-violet and
real blue. The ultra-violet appears as a tiny tint to
the red and blue on visible lightmap.