GameMenu: require confirmation when aborting or restarting a level
Add an "Are you sure?" prompt when aborting or restarting a level. Also, add an
option `confirmation_dialog` (false by default) to toggle on and off this
behavior.

Remove old code that implemented a similar behavior for the "Quit" button in the
main menu, but could never be reached by the user.

Closes SuperTux/supertux#599 (""Restart level" button really needs an "Are you
sure?" prompt")
Closes SuperTux/supertux#724 ("GameMenu: add a submenu for confirming level
restarts")
Merges SuperTux/supertux#843 ("Add confirmation dialog on level reset and
exit.")