First part of fixing AmbientSound bug.
- This modifies how the AmbientSound distance is calculated, since the previous distance was not properly calculated.
It squared the distance values and compared squares and non-squared values, resulting in ambient sound boxes of fixed with / height to not function properly.
This also changes the distance to be based on Tux's position instead of the camera's position.

- Currently, only one ambient sound will play. It appears that it's always either the first ambient-sound listed in the .stl, or the first ambient-sound without a fixed width / height (level ambient sound).

- Line 180 contains a comment block of notes for how the rest may be gone about being fixed.

Addresses and is working on fixing https://github.com/SuperTux/supertux/issues/806.

[ci skip]