Rewrote Rectf to use Vector and Size instead of two Vector
This is to ensure that the Rectf's width and height stay constante,
when two Vectors are used the width and height will vary slightly when
the Rectf is moved around due to floating point rounding errors.

Furthermore this way to represent a Rect seems much more common and is
used in Qt, Java, Gtk, SDL among many others.