Add an optional gain parameter to SoundManager::play
It can be used for example to make loud explosions.
The default gain is 0.5f and valid values are in [0, 1]
because if the Sound Volume is set to 100% in the settings, a gain value of more than 1 is reduced to 1.
Since the sounds are now played half as loud with default gain, the default Sound Volume setting is now 100%