Change worldmap camera panning
The speed does no longer depend on the framerate.
The camera position is smoothly interpolated with a cosine function instead of linear interpolation.
The panning time is mostly direct proportional to the distance between start and end position.
If the distance is very high, this time is reduced to a maximum panning time.