Created solid walls in igloos
When player gets to the end of a level with an ice cube in his hands is gets released and passes through back wall of an igloo like a ghost. This PR adds an invisible wall to each igloo in Antarctica world, so that ice cubes can bounce back and get killed on immortal Tux.