Crash fix: handle dragging a selection rectangle upwards or to the left correctly.
Previously, this created a Rectf instance with negative size, crashing
the filled rectangle drawing routine.

I previously considered rather making the drawing routine support this
case, but then I started to wonder how to define the rounded rectangle
radius in that case... and then I saw the pre-existing assert in the
other Rectf constructor and decided that the intention was to not allow
this in the first place. So, dragging code fixed instead.