Improved code quality in fade_helper.cpp
(Repeating an internal discussion here) This kind of change is recommended when the body of the if statement is considered a special case of some sort. If both cases are considered equivalent in importance and complementary (say, an easing function that uses a certain body for values [0, 0.5[ and another for [0.5, 1]), then keeping the else is recommended so that the code *looks* like it gives equal importance to both parts.