Fixed weird crash in the editor
Prior to this commit, playtesting from the editor with two players leaves a dangling moment for a very short moment when the level intro black screen is created. This may or may not result in a segmentation fault. I admit I'm not entirely sure how the lifetime is managed there, but this fix seems to resolve the issue. It does, however, block a lot of stuff in Sector::activate() and I didn't check everything.