Rocks now apply friction when on the ground
Prior to this commit, a rock falling on the ground while in a wind area that affects objects would instantly loose all its velocity upon hitting the ground, then it would slowly accelerate in the direction of the wind and would keep its velocity as long as it stayed in contact with the ground. With this commit, while on the ground, rocks will loose a certain amount of their velocity each frame.