Disabled option to change owls' carried object
The feature was excessively broken and resulted in crashes or excessively weird behavior. Since it hasn't received any attention lately, I disabled it so people don't accidentally use a broken feature (and because it feels unclean to have an unpolished feature available). I did not remove code directly, to make it easier it someone, someday, wants to fix it and re-enable it.