Remove Community Island from the game
Since a big overhaul for the Bonus Islands is coming, I see no sense or purpose for keeping those levels in the game. There should only be the Bonus Islands and the three Holiday/Retro-based worlds, for the latter because they were created solely by team members. An Add-On could be created for the Community Island instead. So only high-quality content will be kept.