ST OST - add "March of the Malevolent" By Servalot
part of my ST OST remake. to feature in the first level of the corrupted forest.

After Tux witnesses the destruction laying ahead of him, he delves into a land of mystery but also familiarity. This is delivered by the "corrupted tree" theme which converts, what will be, the forest theme in a major form to the harmonic minor/aeolian equivalent. After getting somewhat settled into this new environment, the music transforms to reveal the corrupted tree theme heard in the previous cutscene. hints of nolok's theme are also apparent, to give a sense of omnipresence but also reinforce the role of the true perpetrator in corrupting the forest. The music fades to loop.

Composers: Servalot, Wansti (original forest/'ghost' forest theme)
Arrangement: Servalot

Licence CC-BY-SA 4.0

[ci skip]