Refactor undo/redo system to track separate object changes
UndoManager, previously used by the Editor, has been removed and the undo/redo
functionality has been migrated to GameObjectManager, where the approach taken
to save and reproduce object changes is to track each GameObject's data separately.

The new object tracking system can only be enabled inside Sectors, which take place in the
Editor, and can also manually be turned on/off by the user. The maximum amount of saved
changes in the undo stack can also be changed by the user. The undo stack is cleaned up
when pushing to it and when setting a new maximum limit to undo stack changes.

The virtual `save_state()` and `check_state()` GameObject functions can be overriden,
meaning that markers or more complex objects can also save other objects' states,
so they can be tracked.