Improvements to Dart and DartTrap (#2554)
Removes all hard-coded coordinates and offsets in darttrap.cpp, now the darts will always spawn in the centre of the hitbox's height (width for up and down darts). This makes it much more intuitive to create custom sprites for DartTrap. The default sprite is adjusted for this in a way that won't affect existing levels.
It also makes it possible to change the sprite of the dispensed Dart from DartTrap's menu in the editor.
Darts and DartTraps now also support UP and DOWN directions, in addition to LEFT and RIGHT.
This PR also cleans up darttrap.cpp/hpp according to the guidelines, namely adding the m_ prefix to member variables.