Fix missing (or rather, not loading) editor sprite actions (#2590)
Right now, some badguys in the editor show the wrong sprite action (mostly for Direction::AUTO, but fish with "swim" actions for all directions). There was also a bug that the editor checked m_dir instead of m_start_dir for the badguys with Direction::AUTO. (m_dir is the current direction when the level is running, m_start_dir is the direction saved in the .stl level file that the enemy starts in).
Some would show facing in the wrong direction, some would show the wrong sprite action (eg. woken up sleeping spiky instead of sleeping, burning toads, etc.)

This fixes that.