Fix corrupted crusher not spawning roots when hit.top (#2584)
When the corrupted Crusher was added, for some reason, it was only programmed to spawn CrusherRoots when hitting bottom, left or right, but not top. This prevented it from spawning the roots during levelflip. The roots didn't even have a sprite action for that direction.

Also, CRLF→LF in corrupted Crusher and CrusherRoot sprite files.