Fix camera zooming in "manual" mode
Camera zooming in "manual" mode now requires a target center position, and is properly managed.
The camera progresses to the target position simultaneously with the zooming progression.

Target center position can be provided for zooming on other modes, but does not serve a purpose.

All zooming-related scripting functions have new corresponding functions with `_target` prefixes in their name, which allow for setting the target center position. The usage of those is required on "manual" mode.

Fixes #2405.