Fix camera zooming in "manual" mode (#2607)
Camera zooming in "manual" mode now accepts an anchor point, and is properly managed. The camera progresses to the anchor point, relative to the current position, simultaneously with the zooming progression.

Target anchor point can be provided for zooming on other modes, but does not serve a purpose.

All zooming-related scripting functions have new corresponding functions with _anchor prefixes in their name, which allow for setting the target anchor point. If they are not used for zooming on "manual" mode, the default anchor point (ANCHOR_MIDDLE) will be used.

Fixes #2405.