Improvements to object type support, object versioning (#2542)
Object, utilizing object types to have their sprites changed, now change their default sprite, according to their type. This stops the game from having to save the object type's sprite to the level file, when it's supposed to be the default.

Also features code improvements, like the initial `on_type_change` call for `GameObject`s being moved to the `parse_type` function, and changing an object's sprite on type change being moved to `MovingSprite::on_type_change`.

Another major feature, included in this PR, is `GameObject` versioning. It aims to provide a simple way to deprecate features in an object, which would allow for retaining proper functionality in old levels, whilst providing new functionality for new levels. Object versioning is supposed to be used, when introducing changes to objects, which are incompatible with the previous way of functioning of the object.

Fixes #2502.