Fix editor world conflicts when browsing levels from a different world (#2654)
Previously, `EditorLevelSelectMenu`, when provided a world, set it as the current editor world. This, however, caused world conflcits when browsing levels from different from the one of the opened level worlds. Saving the opened level would then save it in the folder of the other world, which was browsed earlier, possibly overriding levels with the same filename from that other world.

`EditorLevelSelectMenu` now preserves the provided world, until it's time to open a level, where it sets its preserved world as the editor world, before setting the chosen level.

Additionally fixes a bug with deleting levels from other worlds, which have the same filename as the current editor level, which was also caused by world conflicts.

Fixes #2652.