Profile manager, general profile info
Adds a `ProfileManager`, which provides `Profile` objects, containing general data about a profile. The manager creates the profile, if it doesn't exist, so it always returns a valid `Profile` object.

Currently, the `Profile` object contains only the `name` property to indicate the displayed name for the profile on the profile menu.

This object is to be used for more properties in the future. My next pull request is going to require preserving the last world, played on a profile, so this is required.