Uncomment `EditorToolbox` object display name warning
The warning was commented out, because prior to commit 24e36861a3b2efc774c82188064f91dc348a4936, the node marker was a part of the objects, so a warning would display every time it's hovered over.

Since the node marker is now displayed as a tool, the warning can be uncommented.