Undo tracking for multiple objects, no stack clear on save
The undo/redo stack system of `GameObjectManager` now allows for tracking multiple objects, via object change stacks. This has been applied both to the sector resizing and tile conversion features.

Additionally, the stacks are no longer cleared on level save. Instead, the unique ID of the last change in the undo stack is saved, and if it's equal to the unique ID of the last change in the undo stack, it indicates that no changes have been performed in a sector since the last save.