Editor: Remove the 1D OKLab color picker
There are reports that the one-dimensional OKLab color selection is unpopular and difficult to use,
so we can remove it to simplify the code and GUI.

This commit removes ItemColorChannelOKLab and most of the OKLab color space code.
Since the two-dimensional color picker uses some of the OKLab code,
it is not removed completely.