Editor zooming (#2813)
By holding CTRL and scrolling with the mouse, the camera in the editor can now be zoomed in and out. There is a set zoom limit: `0.5` - `3.0`.

When zooming in, the camera position will progressively focus on the position of the mouse, relative to the screen center.

Additionally, "src/editor/util.*" has been removed, since it only provided an unneeded `draw_tile` wrapper function.

Mentioned in #2421.