Add option for frame prediction (camera interpolation)
This moves the camera with sub-game-step precision, which makes camera
motion less jittery at frame rates larger than the 64 fps of the game
logic. This is not as good as increasing the game logic rate, but doing
so would increase CPU load (which is bad for battery life and may make
the game unplayable on older machines) and would require careful physics
rework.