Add feature to load/unload badguys in groups
This adds a 'load-group' property to badguys, which ensures that
multiple badguys tagged with the same load-group string will be loaded
and unloaded together.

In particular, this makes it possible to synchronize flying snowballs
and jumpies -- and keep that synchronization, even when the player
leaves the area and comes back. Doing so makes the level appear more
real and prevents players from making parts of a level easier by
carefully walking toward them and desynchronizing the badguys (or
harder if they have bad luck.)