Corrupted version of Swimming Fish & Chasing Fish (#2955)
Made these sprites a while ago, but seemingly forgot to add them...

Right now all they are planned to be is a re-skin for the Swimming Fish and Chasing Fish badguy, without any behavior changes as of now. I couldn't quite decided on anything.

However, to not postpone their inclusion any longer, I'd say we simply add them as an object types for the mentioned entities and add worry about a potential behavior change later.