Incorporate certain object dynamics (#2846)
Original description:

"For my 100th PR, I decided to fix some old issues I had involving the dynamics between some objects.  Included in this PR are the following:
- Rocks no longer hurt Tux and bounce off of him, making rock puzzles more fair.
- Rocks now interact as expected with slick ice, water, bumpers, and trampolines.
- Enemies now interact as expected with water, bumpers, and trampolines. 
- Objects oriented to the side (switches, darttraps, bumpers, stalactites, yeti stalactites, pushbuttons, ispys, and shards) now properly stick to moving objects which look like they can be stuck to (moving tilemaps, platforms, crushers and fallblocks). [Fixes #1782]

Note: I was hoping to incorporate some more general method for point 4, where I didn't have to specify every single interaction and I could just label some objects as "sticky" and "sticked" objects, but this proved to be difficult as I don't think the game is currently capable of describing objects in such a way that I could do that, and I did not want to spend this PR refactoring some part of the game just to include that little thing.  Each of the different possible interactions tends to have to some slight variation from one another, and there aren't very many of these possible interactions, anyway.  

Fixes #2909 Fixes #2912 "

Commits:

* Rock updates [read desc]

- Rocks now interact with bumpers, water, slick ice, and trampolines.
- Rocks no longer hurt Tux.

* Badguy interactions [read desc]

- Now badguys can interact with bumpers, trampolines, and water.

* Some object sticking dynamics

* Add stickiness as an option

* Fix a rock + bumper + moving platform bug [read desc]

Fixed a bug where if you threw a rock at a bumper on a moving platform it would reset to original position.

* Add sticking to ispy, button, and vertical switches

* Add stalactite to dynamics

* Yeti

* Fix some bugs, un-dupe some code

* StickyBadguys - BUT THEY DISAPPEAR FOR SOME REASON???

* Fix the thing

* Fix some issues

* Code fixes

* Update bumber settings

* Update sticky obj settings

* stationary trampolines stay where they are

* Lower bullet gravity underwater

* Leaves float on water

* Fix sticky lag, add crusher sticking

* Improve leaf water floating

* More updates to water mechanics

* Stuck shard keeps same angle

* More fixes

* Frozen enemies float to the top of water

* Revert "Frozen enemies float to the top of water"

This reverts commit f7beaa0a67132a15d9a09a46b15021d1deddd800.

* I give up

* : |

* Finally

* slower

* Frozen enemies bounce off player like rocks

* Fix my reported rock bugs

* Code cual of tea

* Reduce code duplication in "sticky_object.cpp"

* Update badguy.cpp

---------

Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>