Do not check for "main" sector in `Level::initialize()`
This was originally added, so all players for a level were all added to the initial "main" sector directly.

This leads to the game closing due to the unexpected exception in the editor, when trying to open a level without a "main" sector, forcing the user to have to fix that with a text editor.

Doing this is not needed and all players can simply be added to the first sector of a level, from where they will be moved to the "main" sector on its activation, if needed.