Fix dispenser bug (#2990)
Original PR description: "Fixes #2837

This PR makes dispensers active in terms of being an active/inactive badguy always. This is beneficial for many reasons:

Dispensers don't really function as badguys anyway, moreso as other objects, which are always active.
It fixes the bug where dispensers scripting would break if you went away from the dispenser and then came back.
I believe it fixes an unreported issue where gravity-obeying dispensers would fall into the void if they were placed on moving platforms because the dispenser would load below/inside/off the moving platform because the platform was moving and the dispenser wasn't.
However, I know this could be considered kind of a cheap/non-optimal solution to the bug. So if anyone has any better suggestions feel free to take a look at it."