Merge Bomb with MrBomb (#3002)
This PR internally rewrites MrBomb to be more like GoldBomb. It doesn't add any considerable features.

It also removes lots of code duplication. The Bomb class is now obsolete.

As a result, the death script stays consistent between MrBomb and GoldBomb.

This PR also fixes a bug with gold bomb, in which getting out of the safe distance and going back in would freeze gold bomb in the same sprite action until another state change.

Co-authored-by: MatusGuy <MatusGuy@users.noreply.github.com>
Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>