Revert "Generate builds with `RelWithDebInfo` instead of `Release` (#2963)"
This reverts commit 7302e4b02dd671c7637f160ed2dd4f3326212590.

RelWithDebInfo include debug data separately from the executable, which needs
a debugger to be interpreted properly. The debug data does not affect stack
traces generated by the program itself, but increases the executable by more
than 100Mb, which is wasted if the user does not intent to run the exectuable
inside a debugger.