Smooth badguy and other object motion when frame prediction is on (#3044)
This commit extrapolates the drawn position for badguys and some other
moving objects, making their motion look smoother when both the frame
rate is high and frame prediction is on. This is already being done for
Tux, but that is not enough to always produce the illusion of smooth
motion. Without this commit one can at high frame rates see choppy
motion on carried objects when Tux is running, and on quickly moving
badguys, like MrIceBlock.