Editor: Support multiple autotilesets per tile
Tiles can now be present in multiple autotilesets in the same tileset. They would be autotileable with all (max. 10) autotilesets they are included in.

The editor allows for pressing Ctrl+1, Ctrl+2, etc... up until Ctrl+9 to allow for autotiling a tile with additional 9 autotilesets. This is visually indicated in the autotile help strings.